+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+
+-   Command to reset user live snippets file
+-   Command to compare user live snippets file to default
+
 ## [0.3.4] — 2019-11-02
 
 ### Added